+++ src/components/input/Input.stories.ts
@@ -2,7 +2,9 @@ import { boolean, text, select } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
 import Vue, { PropType } from 'vue';
 import WvuiInput from './Input.vue';
+import WvuiButton from '../button/Button.vue';
 import { InputType } from './InputType';
+import { wvuiIconSearch, wvuiIconInfo } from '@/themes/icons';
 import './Input.stories.less';
 
 export default {
@@ -20,7 +22,7 @@ export const configurable = (): Vue.Component =>
 				type: String as PropType<keyof typeof InputType>,
 				default: select( 'Input Type', Object.keys( InputType ), 'Search' )
 			},
-			placeholder: { type: String, default: text( 'Placeholder', 'Search...' ) }
+			placeholder: { type: String, default: text( 'Placeholder', 'Search…' ) }
 		},
 		data() {
 			return {
@@ -34,21 +36,21 @@ export const configurable = (): Vue.Component =>
 			blur: action( 'blur' )
 		},
 		template: `
-		<div class="sb-input-preview">
+		<div class="sb-input">
 			<wvui-input
 				:placeholder="placeholder"
-				:type="InputType[ type ]"
 				:disabled="disabled"
+				:type="InputType[type]"
 				@input="input"
 				@change="change"
-				@focus="focus" 
+				@focus="focus"
 				@blur="blur"
 			/>
 		</div>
 		`
 	} );
 
-export const withIcon = (): Vue.Component =>
+export const withStartIcon = (): Vue.Component =>
 	Vue.extend( {
 		components: { WvuiInput },
 		props: {
@@ -56,23 +58,24 @@ export const withIcon = (): Vue.Component =>
 		},
 		data() {
 			return {
-				InputType
+				InputType,
+				searchIcon: wvuiIconSearch
 			};
 		},
 
 		template: `
-		<div class="sb-input-preview">
+		<div class="sb-input">
 			<wvui-input
-				placeholder="Search..."
+				placeholder="Search…"
 				:type="InputType.Search"
-				icon="search"
+				:start-icon="searchIcon"
 				:disabled="disabled"
 			/>
 		</div>
 	`
 	} );
 
-export const withIndicator = (): Vue.Component =>
+export const withEndIcon = (): Vue.Component =>
 	Vue.extend( {
 		components: { WvuiInput },
 		props: {
@@ -80,16 +83,16 @@ export const withIndicator = (): Vue.Component =>
 		},
 		data() {
 			return {
-				InputType
+				InputType,
+				endIcon: wvuiIconInfo
 			};
 		},
-
 		template: `
-		<div class="sb-input-preview">
+		<div class="sb-input">
 			<wvui-input
-				placeholder="Search..."
+				placeholder="Search…"
 				:type="InputType.Search"
-				indicator="clear"
+				:end-icon="endIcon"
 				:disabled="disabled"
 			/>
 		</div>
@@ -111,15 +114,71 @@ export const withClearAction = (): Vue.Component =>
 			input: action( 'input' )
 		},
 		template: `
-		<div class="sb-input-preview">
+		<div class="sb-input">
 			<wvui-input
-				placeholder="Search..."
+				placeholder="Type something…"
 				:type="InputType.Search"
 				:clearable="true"
 				:disabled="disabled"
 				value="Some value"
 				@input="input"
 			/>
+		</div>
+	`
+	} );
+
+export const withButton = (): Vue.Component =>
+	Vue.extend( {
+		components: { WvuiInput, WvuiButton },
+		props: {
+			disabled: { type: Boolean, default: boolean( 'Disabled', false ) }
+		},
+		template: `
+		<div class="sb-input sb-input--has-button">
+			<wvui-input
+				placeholder="Search…"
+				:disabled="disabled"
+			/>
+			<wvui-button :disabled="disabled">Search</wvui-button>
+		</div>
+	`
+	} );
+
+const searchLanguageMap = {
+	English: 'Search',
+	Russian: 'Искать',
+	Vietnamese: 'Tìm kiếm',
+	Japanese: '探す',
+	Greek: 'Αναζήτηση',
+	Swedish: 'Söka',
+	Mazandeerani: 'جستجو کردن'
+};
+
+export const wikipediaSearchInput = (): Vue.Component =>
+	Vue.extend( {
+		components: { WvuiInput, WvuiButton },
+		props: {
+			disabled: { type: Boolean, default: boolean( 'Disabled', false ) },
+			buttonLabel: {
+				type: String,
+				default: select( 'Label language', searchLanguageMap, 'Search' )
+			}
+		},
+		data() {
+			return {
+				startIcon: wvuiIconSearch
+			};
+		},
+		template: `
+		<div class="sb-input sb-input--has-button">
+			<wvui-input
+				placeholder="Search…"
+				icon="search"
+				:disabled="disabled"
+				:start-icon="startIcon"
+				:clearable="true"
+			/>
+			<wvui-button :disabled="disabled">{{ buttonLabel }}</wvui-button>
 		</div>
 	`
 	} );