+++ src/components/typeahead-search/TypeaheadSearch.snap.ts
@@ -1,0 +1,87 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`matches a snapshot 1`] = `
+<div
+  aria-expanded="true"
+  aria-haspopup="listbox"
+  aria-owns="wvui-typeahead-search__suggestions"
+  class="wvui-typeahead-search"
+  role="combobox"
+>
+  <form
+    action="https://wikipedia.org/search-redirect.php"
+    class="wvui-search-form"
+    id="wvui-typeahead-search__search"
+  >
+    <input
+      name="family"
+      type="hidden"
+      value="wikipedia"
+    />
+     
+    <input
+      name="language"
+      type="hidden"
+      value="en"
+    />
+     
+    <div
+      class="wvui-input wvui-input--has-start-icon"
+    >
+      <input
+        accesskey="f"
+        aria-autocomplete="list"
+        aria-controls="wvui-typeahead-search__suggestions"
+        aria-label="Search Wikipedia"
+        autocapitalize="off"
+        autocomplete="off"
+        class="wvui-input__input"
+        dir="auto"
+        id="wvui-typeahead-search__input"
+        name="search"
+        placeholder="Search Wikipedia"
+        title="Search Wikipedia [Alt+Shift+f]"
+        type="search"
+      />
+       
+      <span
+        class="wvui-input__start-icon"
+      >
+        <span
+          class="wvui-icon"
+        >
+          <svg
+            aria-hidden="true"
+            height="20"
+            viewBox="0 0 20 20"
+            width="20"
+            xmlns="http://www.w3.org/2000/svg"
+          >
+            <!---->
+             
+            <g
+              fill="#000"
+            >
+              <path
+                d="M7.5 13c3.04 0 5.5-2.46 5.5-5.5S10.54 2 7.5 2 2 4.46 2 7.5 4.46 13 7.5 13zm4.55.46A7.432 7.432 0 017.5 15C3.36 15 0 11.64 0 7.5S3.36 0 7.5 0C11.64 0 15 3.36 15 7.5c0 1.71-.57 3.29-1.54 4.55l6.49 6.49-1.41 1.41-6.49-6.49z"
+              />
+            </g>
+          </svg>
+        </span>
+      </span>
+       
+      <!---->
+    </div>
+     
+    <button
+      class="wvui-button wvui-button--default wvui-button--framed"
+    >
+      
+			Search
+		
+    </button>
+  </form>
+   
+  <!---->
+</div>
+`;