+++ src/components/typeahead-search/TypeaheadSearch.test.ts
@@ -1,0 +1,8 @@
+import { mount } from '@vue/test-utils';
+import WvuiTypeaheadSearch from './TypeaheadSearch.vue';
+
+it( 'matches a snapshot', () => {
+	const wrapper = mount( WvuiTypeaheadSearch );
+
+	expect( wrapper.element ).toMatchSnapshot();
+} );