+++ changelog.md
@@ -4,6 +4,12 @@ Versions and bullets are arranged chronologically from latest to oldest.
 
 ## v0.0.2 (unreleased)
 
+-   [component][typeahead-suggestion] Add keyboard support fot suggestions list
+-   [component][typeahead-suggestion-title] Add highlighting component
+-   [component][typeahead-suggestion] Add initial styles and props
+-   [build] Split icons into a distinct entry
+-   [component][input] Add button control for input
+-   [component][icon] Add component, styles, and icon data
 -   [component][typeahead-search] Add network client
 -   [docs][dev] Document way to clear Jest cache
 -   [build][dev] Add production and pre-release scripts